requirements.txt
pandas
numpy
requests
python-dotenv
scikit-learn
xgboost
joblib
jinja2
yfinance
